Add files via upload
This project is about Prediction using the Supervised Machine Learning technique. Here I create a simple linear regression model to predict the percentage score of a student based on the no. of study hours. I used the R programme in this project.
This task was given in Graduate Rotational Internship Program by The Sparks Foundation.